Removed utility functions from main
